Refactor  Double prices to BigDecimal in Plate class

+++ src/main/resources/db/migration/V2__Create_Plate_Table.sql
@@ -3,6 +3,6 @@ CREATE TABLE plate (
   name VARCHAR(255) NOT NULL,
   available_quantity INT,
   photo VARCHAR(255),
-  price DOUBLE,
+   price DECIMAL(10, 2),
   is_deleted TINYINT(1) DEFAULT FALSE
 );
